@@ -1,13 +1,12 @@
 ﻿using System.Drawing;
 
-namespace BlueHarvest.Core.Extensions
-{
-   public static class RectangleExtensions
-   {
-      public static void CenterOn(this ref Rectangle rect, Point point) =>
-         rect.CenterOn(point.X, point.Y);
+namespace BlueHarvest.Core.Extensions;
 
-      public static void CenterOn(this ref Rectangle rect, int x, int y) =>
-         rect.Location = new Point(x - (rect.Width / 2), y - (rect.Height / 2));
-   }
+public static class RectangleExtensions
+{
+   public static void CenterOn(this ref Rectangle rect, Point point) =>
+      rect.CenterOn(point.X, point.Y);
+
+   public static void CenterOn(this ref Rectangle rect, int x, int y) =>
+      rect.Location = new Point(x - (rect.Width / 2), y - (rect.Height / 2));
 }
