@@ -1,16 +1,14 @@
-﻿using System;
+﻿namespace BlueHarvest.Core.Extensions;
 
-namespace BlueHarvest.Core.Extensions
+public static class DoubleExtensions
 {
-   public static class DoubleExtensions
+   public static bool EqualsToPrecision(this double left, double right, double precision = 0.0001)
    {
-      public static bool EqualsToPrecision(this double left, double right, double precision = 0.0001)
-      {
 #if DEBUG
-         var test = Math.Abs(left - right);
+      var test = Math.Abs(left - right);
 #endif
 
-         return Math.Abs(left - right) < precision;
-      }
+      return Math.Abs(left - right) < precision;
    }
 }
+
