@@ -1,10 +1,7 @@
-﻿using System;
+﻿namespace BlueHarvest.Core.Extensions;
 
-namespace BlueHarvest.Core.Extensions
+public static class MathExtensions
 {
-   public static class MathExtensions
-   {
-      public static double ToRadians(this double degrees)
-         => (Math.PI / 180) * degrees;
-   }
+   public static double ToRadians(this double degrees)
+      => (Math.PI / 180) * degrees;
 }
